[Refactor] Extract http4sInputMatcher
